requirements.txt
tensorflow>=2.7.0
numpy>=1.18.0
pandas>=1.0.0
matplotlib>=3.2.0
scikit-learn>=0.24.0
yfinance>=0.1.54
pyyaml>=5.4
pytest>=6.0